Created the SecondFactorTypeService
This service will be used to retrieve information about SecondFactor
types. The SecondFactorType (SFT) was dumbed down to no longer track the
LOA level of the SFT. This configuration is now tracked partly in YML
(for the generic SAML second factor types (gssf)) and partly in the
service itself.

A new Value object was introduced to encapsulate the GssfConig.

Tests where added and moved to keep the same and actually better code
coverage.