Implement error page redesign
In order for the exception controller to be able to show specific
messages depending on the exception type, it needs access to the
original exception object instead of the FlattenException which the
symfony exception listener passes on to the controller. In order to
make this work, the stepup bundle now overrides the default exception
controller.